[refactor] page base components
