+++ client/src/app/app-routing.module.ts
@@ -2,8 +2,8 @@ import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 
 // page components
-import { DashboardComponent } from './components/page/dashboard/dashboard.component';
-import { ClusterComponent } from './components/page/cluster/cluster.component';
+import { DashboardComponent } from './component/dashboard/dashboard.component';
+import { ClusterComponent } from './component/cluster/cluster.component';
 
 const routes: Routes = [
   { path: '', component: DashboardComponent },
