+++ client/src/app/component/cluster/cluster.component.html
@@ -1,6 +1,6 @@
 <div>
   <app-header></app-header>
-  <app-navbar [product]="product">
+  <app-sidenav [product]="product">
     <div class="scroll-content">
       <app-topic-path></app-topic-path>
 
@@ -25,5 +25,5 @@
         </mat-card>
       </div>
     </div>
-  </app-navbar>
+  </app-sidenav>
 </div>
