+++ client/src/app/component/dashboard/dashboard.component.html
@@ -1,6 +1,6 @@
 <div>
   <app-header></app-header>
-  <app-navbar [product]="product">
+  <app-sidenav [product]="product">
     <div class="scroll-content">
       <app-topic-path></app-topic-path>
 
@@ -14,5 +14,5 @@
         </div>
       </div>
     </div>
-  </app-navbar>
+  </app-sidenav>
 </div>
