+++ client/src/app/component/dashboard/product-card/product-card.component.ts
