+++ client/src/app/component/dashboard/product-version-chip/product-version-chip.component.ts
