+++ client/src/app/shared/shared.module.ts
@@ -28,12 +28,9 @@ import { FlexLayoutModule } from '@angular/flex-layout';
 import { NgbModalModule } from '@ng-bootstrap/ng-bootstrap';
 
 // components
-import { HeaderComponent } from './components/header/header.component';
-import { NavbarComponent } from './components/navbar/navbar.component';
-import { TopicPathComponent } from './components/topic-path/topic-path.component';
 
 @NgModule({
-  declarations: [HeaderComponent, NavbarComponent, TopicPathComponent],
+  declarations: [],
   imports: [
     FormsModule,
     ReactiveFormsModule,
@@ -102,11 +99,6 @@ import { TopicPathComponent } from './components/topic-path/topic-path.component
     ScrollingModule,
     FlexLayoutModule,
     NgbModalModule,
-
-    // components
-    HeaderComponent,
-    NavbarComponent,
-    TopicPathComponent,
   ],
 })
 export class SharedModule {}
